requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.1
scikit-learn==1.3.0
tensorflow==2.13.0
transformers==4.33.2
torch==2.0.1
spacy==3.6.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
textblob==0.17.1


